Simplify commands, fix queue consts
