+++ bin/queue-count-failed
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-php /usr/src/code/app/cli.php queue-count-failed $@
+php /usr/src/code/app/cli.php queue-count --type=failed $@
