+++ bin/queue-count-processing
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-php /usr/src/code/app/cli.php queue-count-processing $@
+php /usr/src/code/app/cli.php queue-count --type=processing $@
