+++ bin/queue-count-success
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-php /usr/src/code/app/cli.php queue-count-success $@
+php /usr/src/code/app/cli.php queue-count --type=success $@
